@@ -1,6 +1,18 @@
 require_relative 'boot'
 
-require 'rails/all'
+require 'rails'
+%w(
+  active_record/railtie
+  action_controller/railtie
+  action_view/railtie
+  rails/test_unit/railtie
+  sprockets/railtie
+).each do |railtie|
+  begin
+    require railtie
+  rescue LoadError
+  end
+end
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
@@ -9,7 +21,7 @@ Bundler.require(*Rails.groups)
 module Contribulator
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
-    config.load_defaults 5.1
+    config.load_defaults 6.1
 
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
